+++ src/Components/About/index.js
@@ -1,11 +1,13 @@
 import React from 'react'
-import Title from '../Titles/Titles'
+import Title from '../Titles'
 import AboutOng from './AboutOng'
 
 const About = () => {
   return (
     <>
-        <Title text='Nosotros' />
+        <Title imageUrl="">
+          Nosotros
+        </Title>
         <AboutOng />
     </>
   )
